refactor(userid): jk move it back to telemetry
