--- internal/userid/userid.go
@@ -1,56 +1,0 @@
-// Copyright 2026 DataRobot, Inc. and its affiliates.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package userid
-
-import (
-	"context"
-	"errors"
-
-	"github.com/datarobot/cli/internal/config"
-	"github.com/datarobot/cli/internal/drapi"
-)
-
-// AccountInfo represents the response from GET /api/v2/account/info/.
-type AccountInfo struct {
-	UID       string `json:"uid"`
-	Email     string `json:"email"`
-	FirstName string `json:"firstName"`
-	LastName  string `json:"lastName"`
-	TenantID  string `json:"tenantId"`
-	OrgID     string `json:"orgId"`
-}
-
-// GetUserID fetches the DataRobot user uid from GET /api/v2/account/info/.
-// It returns the uid string on success, or ("", error) on non-200 status,
-// empty uid, or network failure.
-func GetUserID(ctx context.Context) (string, error) {
-	url, err := config.GetEndpointURL("/api/v2/account/info/")
-	if err != nil {
-		return "", err
-	}
-
-	var info AccountInfo
-
-	//nolint:contextcheck // GetJSON does not yet accept context; ctx is reserved for future use
-	if err := drapi.GetJSON(url, "", &info); err != nil {
-		return "", err
-	}
-
-	if info.UID == "" {
-		return "", errors.New("empty uid in account info response")
-	}
-
-	return info.UID, nil
-}
